+++ lib/domain/service/my_user.dart
@@ -29,6 +29,7 @@ import '/domain/model/user.dart';
 import '/domain/repository/my_user.dart';
 import '/routes.dart';
 import '/util/log.dart';
+import '/util/obs/rxmap.dart';
 import '/util/web/web_utils.dart';
 import 'auth.dart';
 import 'disposable_service.dart';
@@ -49,6 +50,9 @@ class MyUserService extends DisposableService {
 
   /// Returns the currently authenticated [MyUser].
   Rx<MyUser?> get myUser => _userRepo.myUser;
+
+  /// Returns a reactive map of all the known [MyUser] profiles.
+  RxObsMap<UserId, Rx<MyUser>> get profiles => _userRepo.profiles;
 
   @override
   void onInit() {
@@ -121,8 +125,13 @@ class MyUserService extends DisposableService {
       await WebUtils.protect(() async {
         await _userRepo.updateUserPassword(oldPassword, newPassword);
 
-        // TODO: Replace `force` with something more granular and correct.
-        await _auth.signIn(newPassword, num: myUser.value?.num, force: true);
+        // TODO: Replace `unsafe` with something more granular and correct.
+        await _auth.signIn(
+          newPassword,
+          num: myUser.value?.num,
+          unsafe: true,
+          force: true,
+        );
       });
     });
   }