+++ lib/ui/widget/svg/svgs.dart
@@ -1893,6 +1893,18 @@ class SvgIcons {
     height: 32,
   );
 
+  static const SvgData logout = SvgData(
+    'assets/icons/logout.svg',
+    width: 16.28,
+    height: 19,
+  );
+
+  static const SvgData logoutWhite = SvgData(
+    'assets/icons/logout_white.svg',
+    width: 16.28,
+    height: 19,
+  );
+
   static const SvgData addParticipant = SvgData(
     'assets/icons/add_participant.svg',
     width: 21.48,